- Merge hw-native-sys/simpler main up to a9f3ea9 into the local feature branch
- Resolve the only content conflict in simpler_setup/code_runner.py by keeping the post_run_collect flow and upstream dump-tensor config support
- Preserve the existing host-device share memory work on top of the refreshed upstream base

+++ tests/st/a2a3/tensormap_and_ringbuffer/batch_paged_attention/kernels/orchestration/paged_attention_orch.cpp
@@ -71,7 +71,7 @@ __attribute__((visibility("default"))) void aicpu_orchestration_entry(const Chip
 
     uint64_t scale_value = orch_args.scalar(0);
 
-    uint64_t q_tile = std::min(num_heads, 128UL);
+    uint64_t q_tile = std::min(num_heads, static_cast<uint64_t>(128));
     uint64_t q_loop = (num_heads + q_tile - 1) / q_tile;
     uint64_t elem_size = get_element_size(data_type);
